add default nested cases for client-initialization
